+++ add_cols.py
@@ -16,6 +16,7 @@ c = conn.cursor()
 
 add_col_if_not_exists(c, "users", "premium_expires_at", "DATETIME")
 add_col_if_not_exists(c, "users", "recovery_code_enc", "TEXT")
+add_col_if_not_exists(c, "users", "allow_youtube_import", "BOOLEAN DEFAULT 0")
 
 conn.commit()
 conn.close()